Separate layer pen to own module
